refactor(release-notes): one day rule, and docs that match the code
Review follow-ups on PR #152.

One shared day helper. `utils/dates.ts` now owns both forms of a note's
day - the ISO day the CSV column and the tag+author+day grouping key are
built from, and the human date the card prints. The grouping key used a
raw `createdAt.slice(0, 10)`, so three places carried their own version
of the same rule.

The card date is now read in UTC like the other two. It was local, so a
note stamped late in the UTC evening printed one day on the card and the
day before in its own CSV row, for a reader east of Greenwich. Tested at
both ends of the UTC day.

Dropped the dead `?? note.tag` fallback in the CSV row: TAG_LABELS is a
total `Record<NoteTag, string>`, so the fallback could only ever hide a
type error.

Docs now say what the code does, on two points the review found:
- A divider is a run of three or more rule characters - dashes, and also
  underscores and equals signs, because real files draw the same line
  with all three. ADR-0011 said "mostly dashes". The wide class is kept
  deliberately, and the ADR now records why: a missed divider swallows
  the pages under it into the group above, which is the worse failure.
- The no-Foundation fallback offers every *non-divider* page. ADR-0011
  and CONTEXT.md both said "every page"; the panel notice said it too.
